refactor platform-specific code to platform folder, initial wasm build setup
