+++ src/lib.rs
@@ -1,0 +1,14 @@
+/// lib exports for use as library
+
+mod utils;
+// mod viewer;
+// use crate::viewer::{GltfViewer, CameraOptions};
+
+mod shader;
+mod controls;
+mod importdata;
+// TODO!: adapt Source...
+// mod http_source;
+// use http_source::HttpSource;
+mod platform;
+pub mod render;
